delete old code & move file
