+++ manga_translator/mode/share.py
